Attempt to abstract out each fetch type (requests/playwright/webdriver etc)
